requirements.txt
bresenham==0.2.1
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==9.4.0
Pillow==10.0.1
python-dotenv==1.0.0
Requests==2.31.0
